Fix correct and incorrect guess in run.py
Add docstring to line 1 of run.py.  Remove commented out
code at line 38 and added a print statement statement
inside play_hangman to display what letter was being chosen
as the print statement inside the obtain character was
giving an error.  Added a try statement to line 66. In the
elif statement was referencing character_complete instead of
character changed this to character.  Added print statement
"you loose a life" if the guess is not in the character so
the player knows.  Also fixed all code where lines were too
long in run.py.   In score.py added docstring to the top
of file explaining what score.py holds.  Moved comment to a
new line 10.