Changed constants and fixed errors in score.py.
Constants win, loose, welcome were not being recognised in score.py
so changed them all to uppercase.  In run.py at the top of file
changed to uppercase and down in the body of code also changed the
print statements to uppercase.  In score.py there was over 100 errors
in realtion to "invalid escape '\' and Anomalous backslash in string:
'\'.".  In my ASCII Text banners changed all single backslashes and
removed the strip statements. Also added a double set of backslahes to
hangman introduction as it was not printing hand or leg part right.